refactor(DAL): move PaginationData record into DataTransfer namespace
