+++ NjordinSight.DAL/EntityModelService/Abstractions/IModelReaderService.cs
@@ -1,4 +1,5 @@
-﻿using NjordinSight.UserInterface;
+﻿using NjordinSight.DataTransfer;
+using NjordinSight.UserInterface;
 using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
